refactor(Cards): create wrapper class to hold cards methods
